[Build 619] [Update] Fin de la GUI (non débug) Il manque le key mapping

+++ Assets/Script/FrameWork/Profile.cs
@@ -46,7 +46,7 @@ public class Profile{
 	//Video
 	public bool enableBloom;
 	public bool enableDepthOfField;
-	public float antiAliasing;
+	public int antiAliasing;
 	
 	
 	//Songs
