+++ Assets/Script/Managers/DataManager.cs
@@ -88,7 +88,7 @@ public class DataManager{
 	//Video
 	public bool enableBloom;
 	public bool enableDepthOfField;
-	public float antiAliasing;
+	public int antiAliasing;
 	
 	
 	//OPTION WHEEL
